Implement tests for Image Carousel component and separate test files
